Fixing csleep
generic version of csleep was broken, probably due to type conversion issues, now fixed

+++ libsrc/stdlib/csleep.c
@@ -4,7 +4,7 @@
  *
  *	djm 15/10/2001
  *
- *	$Id: csleep.c,v 1.3 2016-07-02 14:44:33 dom Exp $
+ *	$Id: csleep.c,v 1.3 2016/07/02 14:44:33 dom Exp $
  */
 
 #include <stdlib.h>
@@ -17,7 +17,7 @@
 void  csleep(unsigned int centiseconds)
 {
 	long start = clock();  
-	long per   = (long) centiseconds * CLOCKS_PER_SEC / 100;
+	long per   = ((long) centiseconds * (long) CLOCKS_PER_SEC) / 100L;
 #ifdef __ZX80__
 	gen_tv_field_init(0);
 #endif
